Attempt to change "host" to "client".
This matches the word used in RFC 9663. It's arguably more
correct as well. Plus, it allows us to remove the strange "this
section uses the word host even if the client is a router" text
at the beginning of the host behaviour section.

A few instances continue to use the word "host", when that is
appropriate.